requirements.txt
transformers
datasets
torchvision
pytesseract
python-docx
opencv-python
Pillow
torch
discord
spacy
dotenv